requirements.txt
aiofiles==0.6.0
aiohttp==3.8.3
aiohttp-socks==0.7.1
aiosignal==1.2.0
asttokens==2.0.8
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
cachetools==4.2.4
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
decorator==5.1.1
executing==1.1.0
frozenlist==1.3.1
future==0.18.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==4.12.0
ipython==8.5.0
jedi==0.18.1
jsonschema==3.2.0
jupyter-core==4.11.1
Logbook==1.5.3
Markdown==3.4.1
matplotlib-inline==0.1.6
matrix-nio==0.19.0
multidict==6.0.2
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
peewee==3.15.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-cryptography-fernet-wrapper==1.0.4
python-magic==0.4.27
python-olm==3.1.3
python-socks==2.0.3
simplematrixbotlib==2.7.0
six==1.16.0
stack-data==0.5.1
toml==0.10.2
traitlets==5.4.0
unpaddedbase64==2.1.0
wcwidth==0.2.5
yarl==1.8.1
zipp==3.8.1
